Power gui & rebalance
Stored power now shown on GUI
Reactor heat rebalanced again, a 3x3x3 should now produce modest power
and heat up to circa 600C.
Power is now accumulated as floating point during the reactor loop, to
prevent truncation problems at low values

+++ erogenousbeef/bigreactors/api/HeatPulse.java
@@ -1,11 +1,11 @@
 package erogenousbeef.bigreactors.api;
 
 public class HeatPulse {
-	public int powerProduced;
+	public double powerProduced;
 	public double heatChange;
 	
 	public HeatPulse() {
-		powerProduced = 0;
+		powerProduced = 0.0;
 		heatChange = 0.0;
 	}
 }
